Updated Player-Enemy Collision Algorithm and moved some functions into /utils/collision.ts
